+++ packages/nestjs-util/src/filter/guard/enriched-exception.guard.ts
@@ -1,0 +1,9 @@
+import { EnrichedExceptionError } from '@filter/interface'
+
+export function isEnrichedException (exception: unknown): exception is EnrichedExceptionError {
+  if (exception instanceof EnrichedExceptionError) {
+    return true
+  }
+
+  return false
+}