+++ packages/nestjs-util/src/filter/guard/http-exception.guard.ts
@@ -1,0 +1,9 @@
+import { HttpException } from '@nestjs/common'
+
+export function isHttpException (exception: unknown): exception is HttpException {
+  if (exception instanceof HttpException) {
+    return true
+  }
+
+  return false
+}