+++ packages/nestjs-util/src/filter/index.ts
@@ -1,4 +1,5 @@
+export * from './guard'
+export * from './interface'
 export * from './global-exception.filter'
 export * from './global-sentry-exception.filter'
 export * from './bad-request-exception.filter'
-export * from './exception.interface'