+++ packages/nestjs-util/src/filter/interface/index.ts
@@ -1,0 +1,2 @@
+export * from './class-validator-exception.interface'
+export * from './enriched-exception.interface'